pyproject.toml
[tool.poetry]
name = "Great Expectations avec Disney."
version = "0.1.0"
description = "Validation de la qualité des données avec Great Expectations."
authors = ["Youcef Kharroubi <youcef.kharroubi@soyhuce.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.9.16"
great_expectations = "~0.17.5"
pre-commit = "^2.21.0"
coverage = "^7.0.5"
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
psycopg2-binary = "^2.9.6"
sqlalchemy = "^2.0.19"


[tool.poetry.group.dev.dependencies]
pylint = "^2.15.10"
black = "^22.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit=["tests/*", "data/*", "output/*"]

[tool.black]
line-length = 100

[tool.pylint.design]
# Maximum number of arguments for function / method.
max-args = 6  # Default: 5

# Maximum number of attributes for a class (see R0902).
max-attributes = 15  # Default: 7

# Maximum number of boolean expressions in an if statement (see R0916).
max-bool-expr = 5

# Maximum number of branch for function / method body.
max-branches = 12

# Maximum number of locals for function / method body.
max-locals = 15

# Maximum number of parents for a class (see R0901).
max-parents = 7

# Maximum number of public methods for a class (see R0904).
max-public-methods = 20

# Maximum number of return / yield for function / method body.
max-returns = 6

# Maximum number of statements in function / method body.
max-statements = 50

# Minimum number of public methods for a class (see R0903).
min-public-methods = 0  # Default: 2

[tool.pylint.format]
# Maximum number of characters on a single line.
max-line-length = 100

# Maximum number of lines in a module.
max-module-lines = 1000

[tool.pylint.imports]
allow-wildcard-with-all = true

[tool.pylint."messages control"]
# Only show warnings with the listed confidence levels. Leave empty to show all.
# Valid levels: HIGH, CONTROL_FLOW, INFERENCE, INFERENCE_FAILURE, UNDEFINED.
confidence = ["HIGH", "CONTROL_FLOW", "INFERENCE", "INFERENCE_FAILURE", "UNDEFINED"]

# Disable the message, report, category or checker with the given id(s). You can
# either give multiple identifiers separated by comma (,) or put this option
# multiple times (only on the command line, not in the configuration file where
# it should appear only once). You can also use "--disable=all" to disable
# everything first and then re-enable specific checks. For example, if you want
# to run only the similarities checker, you can use "--disable=all
# --enable=similarities". If you want to run only the classes checker, but have
# no Warning level messages displayed, use "--disable=all --enable=classes
# --disable=W".
disable = [
    # Defaults
    "raw-checker-failed",
    "bad-inline-option",
    "locally-disabled",
    "file-ignored",
    "suppressed-message",
    "useless-suppression",
    "deprecated-pragma",
    "use-symbolic-message-instead",

    # Custom
    # Disabled checkers
    "similarities",

    # Disabled messages
    "redefined-outer-name",
    "import-error",  # Pylint doesn't understand anything to imports, let Pylance handle them.

    # Temporary to avoid too many alerts after adding pylint, these are meant to be reactivated
    # as the project moves forward.
    "missing-function-docstring",
    "missing-class-docstring",
    "missing-module-docstring",
    "logging-fstring-interpolation",
    "logging-not-lazy",
    "fixme",
    "invalid-name",
    "consider-using-enumerate",
    "unspecified-encoding",
    "try-except-raise",
    "raise-missing-from",
    "consider-using-dict-items",
]

# Enable the message, report, category or checker with the given id(s). You can
# either give multiple identifier separated by comma (,) or put this option
# multiple time (only on the command line, not in the configuration file where it
# should appear only once). See also the "--disable" option for examples.
enable = ["c-extension-no-member"]


[tool.pylint.miscellaneous]
# List of note tags to take in consideration, separated by a comma.
notes = ["FIXME", "TODO"]

[tool.pylint.refactoring]
# Maximum number of nested blocks for function / method body
max-nested-blocks = 5

[tool.pylint.reports]

# Activate the evaluation score.
score = true
